+++ src/main/java/dev/demonicrous/furusato/api/service/ServiceStatus.java
@@ -1,0 +1,6 @@
+package dev.demonicrous.furusato.api.service;
+
+public enum ServiceStatus {
+    REGISTERED,
+    AVAILABLE
+}