+++ src/main/java/dev/demonicrous/furusato/api/service/ServiceThreadPolicy.java
@@ -1,0 +1,8 @@
+package dev.demonicrous.furusato.api.service;
+
+public enum ServiceThreadPolicy {
+    THREAD_SAFE,
+    CLIENT_THREAD_ONLY,
+    SERVER_THREAD_ONLY,
+    IMMUTABLE
+}